# Conflicts:
#	smstrikers-port/CMakeLists.txt

+++ smstrikers-port/include/NL/glx/glxTarget.h
@@ -18,6 +18,6 @@ bool glx_GetSharedLock();
 void glx_UnlockSharedMemory();
 void glx_LockSharedMemory();
 u32 glx_GetSharedMemorySize();
-u32 glx_GetSharedMemory();
+uintptr_t glx_GetSharedMemory();
 
 #endif // _GLXTARGET_H_
